Support for sentence interpretation, and minor object modifications
The sentence interpretation was implemented purely out of convenience reasons, allowing to call a whole pick and place plan simply by typing

```lisp
(ex "move pancakemix from counter to sink")
```

And a whole plan will be generated from this, parameterizing all necessary designators. Again, this is a convenience feature for developers and is not intended to support natural language interpretation.